+++ packages/e2e-tests/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- Change e2e-test to test erc20.trasfer, refund, and confidential payments
 - Add check for `OperationProcessed` event
 - Add e2e-test for standalone `SubtreeUpdateServer`
 - Modify e2e-test to use `SubtreeUpdater`